+++ src/palace/manager/metadata_layer/base/frozen.py
