+++ src/palace/manager/metadata_layer/base/mutable.py
@@ -3,13 +3,14 @@ from __future__ import annotations
 from pydantic import BaseModel, ConfigDict
 from sqlalchemy.orm import Session
 
+from palace.manager.core.exceptions import PalaceValueError
 from palace.manager.metadata_layer.identifier import IdentifierData
 from palace.manager.sqlalchemy.model.datasource import DataSource
 from palace.manager.sqlalchemy.model.identifier import Identifier
 from palace.manager.util.log import LoggerMixin
 
 
-class BaseDataContainer(BaseModel, LoggerMixin):
+class BaseMutableData(BaseModel, LoggerMixin):
     model_config = ConfigDict(
         frozen=False,
         validate_assignment=True,
@@ -30,12 +31,12 @@ class BaseDataContainer(BaseModel, LoggerMixin):
         return self._data_source
 
     def load_primary_identifier(self, _db: Session) -> Identifier:
-        """Find the Identifier associated with this circulation information."""
+        """Find the Identifier associated with this data."""
         if self._primary_identifier is None:
             if self.primary_identifier_data:
                 obj, ignore = self.primary_identifier_data.load(_db)
                 self._primary_identifier = obj
                 return obj
             else:
-                raise ValueError("No primary identifier provided!")
+                raise PalaceValueError("No primary identifier provided!")
         return self._primary_identifier
