+++ src/palace/manager/metadata_layer/link.py
@@ -7,7 +7,7 @@ from frozendict import frozendict
 from pydantic import Field, constr, field_validator, model_validator
 from typing_extensions import Self
 
-from palace.manager.metadata_layer.frozen_data import BaseFrozenData
+from palace.manager.metadata_layer.base.frozen import BaseFrozenData
 from palace.manager.sqlalchemy.model.resource import Hyperlink, Representation
 from palace.manager.util.log import LoggerMixin
 from palace.manager.util.pydantic import FrozenDict
