+++ src/palace/manager/metadata_layer/policy/presentation.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing_extensions import Self
 
-from palace.manager.metadata_layer.frozen_data import BaseFrozenData
+from palace.manager.metadata_layer.base.frozen import BaseFrozenData
 
 
 class PresentationCalculationPolicy(BaseFrozenData):
