+++ src/palace/manager/metadata_layer/policy/replacement.py
@@ -4,7 +4,7 @@ from pydantic import Field
 from sqlalchemy.orm import Session
 from typing_extensions import Self, TypedDict, Unpack
 
-from palace.manager.metadata_layer.frozen_data import BaseFrozenData
+from palace.manager.metadata_layer.base.frozen import BaseFrozenData
 from palace.manager.metadata_layer.policy.presentation import (
     PresentationCalculationPolicy,
 )
